marketing-website/pyproject.toml
[project]
name = "mandari-marketing-website"
version = "0.9.0"
description = "Mandari Marketing Website - Wagtail CMS"
requires-python = ">=3.12"
license = {text = "EUPL-1.2"}

dependencies = [
    "django>=6.0,<7.0",
    "wagtail>=6.4,<7.0",
    "dj-database-url>=2.3.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.0",
    "whitenoise>=6.8.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
